+++ src/app/core/guards/admin.guard.ts
@@ -1,0 +1,51 @@
+import { Injectable } from '@angular/core';
+import {
+  CanActivate,
+  ActivatedRouteSnapshot,
+  RouterStateSnapshot,
+  UrlTree,
+  Router,
+  CanLoad,
+  Route,
+  UrlSegment,
+  CanActivateChild
+} from '@angular/router';
+
+import { Observable } from 'rxjs';
+
+import { AuthService } from 'src/app/core/services/auth.service';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AdminGuard implements CanActivate, CanActivateChild, CanLoad {
+  constructor(
+    private auth: AuthService,
+    private router: Router
+  ) {}
+
+  canActivate(
+    next: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+      return this.checkLogin();
+  }
+
+  canActivateChild(
+    childRoute: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): boolean | UrlTree | Observable<boolean | UrlTree> | Promise<boolean | UrlTree> {
+    return this.checkLogin();
+  }
+
+  canLoad(route: Route, segments: UrlSegment[]): boolean | Observable<boolean> | Promise<boolean> {
+    return this.checkLogin() as boolean;
+  }
+
+  private checkLogin() {
+    if (this.auth.isUserAuthorized) {
+      return true;
+    } else {
+      this.router.navigate(['/login']);
+      return false;
+    }
+  }
+}