+++ src/app/core/models/account.ts
@@ -1,6 +1,6 @@
 export class Account {
-	constructor(
-		public login: string,
-		public password: string
-	) {}
+    constructor(
+        public login: string,
+        public password: string
+    ) {}
 }